% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateConfigurations.R
\name{EvaluateConfigurations}
\alias{EvaluateConfigurations}
\title{Evaluate candidate configurations on sets of instances}
\usage{
EvaluateConfigurations(instance.list, N, sampled.instances = numeric(0))
}
\arguments{
\item{instance.list}{list containing all available tuning instances.}

\item{instances.to.eval}{vector containing the indices of the instances to
be used.}

\item{config.list}{list containing all configurations sampled.}

\item{configs.to.eval}{vector containing the indices of the configurations to
be evaluated on the instances listed in
\code{instances.to.eval}.}

\item{target.runner}{name of function used for evaluating the configurations.
See Section \code{Target Runner} of \code{\link[=metatuner]{metatuner()}}
documentation.}
}
\value{
Updated \code{config.list} containing performance evaluation of
\code{configs.to.eval} on \code{instances.to.eval}
}
\description{
Evaluates the candidate configurations listed on \code{configs.to.eval} on the
tuning instances listed in \code{instances.to.eval}. Notice that this routine does
not repeat evaluations - if a configuration \code{i} already has a performance
value for instance \code{j} that evaluation is skipped.
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br}),
Athila Trindade (\email{rochaathila@gmail.com})
}
