% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterRepeatedConfigurations.R
\name{FilterRepeatedConfigurations}
\alias{FilterRepeatedConfigurations}
\title{Filter redundant configurations}
\usage{
FilterRepeatedConfigurations(newconfs, config.list)
}
\arguments{
\item{newconfs}{new configurations returned by \code{\link[=OptimizeModels]{OptimizeModels()}}}

\item{config.list}{archive list}
}
\value{
updated list \code{newconfs} without redundant configurations
}
\description{
Remove repeated configurations from the config.list
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br}),
}
