% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitLassoRidgeModels.R
\name{FitLassoRidgeModels}
\alias{FitLassoRidgeModels}
\title{Fit lasso or ridge regression models}
\usage{
FitLassoRidgeModels(X, Nmodels, model.order, model.type = c("lasso", "ridge"))
}
\arguments{
\item{X}{data frame containing the performance data}

\item{Nmodels}{number of models to generate (1 original plus \code{Nmodels - 1}
perturbations)}

\item{model.order}{order to use for the regression model. Defaults to 2}

\item{model.type}{type of model to fit. See \code{\link[=metatuner]{metatuner()}} for details.}
}
\value{
list containing information about all models fitted
}
\description{
Fits lasso or ridge regression models to the performance data gathered by MetaTuner.
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br}),
Athila Trindade (\email{rochaathila@gmail.com})
}
