% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWeights.R
\name{GetWeights}
\alias{GetWeights}
\title{Calculate weights for regression}
\usage{
GetWeights(X)
}
\arguments{
\item{X}{a matrix containing the (normalized) performance data of
configurations on instances.}
}
\value{
a vector of m values in the (0, 1) interval. Each value is associated
with a configuration, proportionally to the number of instances on which
it has been evaluated.
}
\description{
Calculate weights to be used by Linear and Quantile regression models.
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br}),
Athila Trindade (\email{rochaathila@gmail.com})
}
