% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vectorized_smoof.R
\name{make_vectorized_smoof}
\alias{make_vectorized_smoof}
\title{Make vectorized smoof function}
\usage{
make_vectorized_smoof(prob.name, ...)
}
\arguments{
\item{prob.name}{name of the problem to build}

\item{...}{other parameters passed to each specific function}
}
\description{
Make a vectorized version of test functions available in package "smoof".
}
\details{
This routine builds vectorized versions of the classic test functions
available in package \code{smoof}. Check \code{smoof}'s documentation for details.
}
\examples{
\dontrun{
  library(smoof)
  Rosenbrock <- make_vectorized_smoof(prob.name = "Rosenbrock",
                                 dimensions   = 10)
  Rosenbrock(X = matrix(runif(100), ncol = 10))
}

}
