% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizeModels.R
\name{OptimizeModels}
\alias{OptimizeModels}
\title{Optimize regression models to estimate new configurations}
\usage{
OptimizeModels(parameters, models, optimization.method = "Nelder-Mead",
  ndigits)
}
\arguments{
\item{parameters}{data frame containing information about the tunable
parameters}

\item{models}{list containing models generated by \code{\link[=FitModels]{FitModels()}}.}

\item{optimization.method}{optimization method to be used}

\item{ndigits}{number of significant digits to use for each parameter.}
}
\value{
list containing new candidate configurations found by optimization
}
\description{
Optimize regression models to estimate new configurations
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br}),
Athila Trindade (\email{rochaathila@gmail.com})
}
