% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleInstances.R
\name{SampleInstances}
\alias{SampleInstances}
\title{Sample instances (without replacement) from the set of available tuning
instances}
\usage{
SampleInstances(instance.list, N, sampled.instances = numeric(0))
}
\arguments{
\item{instance.list}{list containing all available tuning instances.}

\item{N}{number of instances to sample.}

\item{sampled.instances}{vector of indices of all instances that have
already been sampled.}
}
\value{
a vector (of size between zero and N) of instance indices
}
\description{
Returns a randomly sampled subset of instances, so that repetitions do not
occur. If the number of available instances in \code{instance.list} is smaller
than \code{N}, it will return a vector containing less than \code{N} indices (it can
also return an empty vector, if all instances have already been sampled)
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br}),
Athila Trindade (\email{rochaathila@gmail.com})
}
