% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SanityCheck.R
\name{SanityCheck}
\alias{SanityCheck}
\title{Check sanity of inputs for metatuner}
\usage{
SanityCheck(myenv)
}
\arguments{
\item{myenv}{list containing input parameters used for \code{\link[=metatuner]{metatuner()}}.}
}
\description{
Check sanity of inputs for metatuner
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br})
}
