% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateConfigurations.R
\name{EvaluateConfigurations}
\alias{EvaluateConfigurations}
\title{Evaluate candidate configurations on sets of instances}
\usage{
EvaluateConfigurations(tuning.instances, instances.to.eval = "all",
  config.list, configs.to.eval = "all", algo.runner,
  summary.function = "median", parameters)
}
\arguments{
\item{tuning.instances}{list containing all available tuning instances.}

\item{instances.to.eval}{vector containing the indices of the instances to
be used.}

\item{config.list}{list vector containing field \code{A} (list of all
configurations sampled so far + additional info) as well
as other information regarding the performance of all
configurations tested and the process of MetaTuner.}

\item{configs.to.eval}{vector containing the indices of the configurations to
be evaluated on the instances listed in
\code{instances.to.eval}.}

\item{algo.runner}{name of function used for evaluating the configurations.
See Section \code{Algorithm Runner} of \code{\link[=metatuner]{metatuner()}}
documentation.}

\item{summary.function}{name of function for aggregating the (scaled)
output values of \code{algo.runner} into a single performance value. Usual
functions include \code{mean} and \code{median}, but in principle any summarizing
function can be used.}

\item{parameters}{data frame containing the parameter names and bound
constraints.}
}
\value{
Updated \code{config.list} containing performance evaluation of
\code{configs.to.eval} on \code{instances.to.eval}, as well as \code{nruns}, the number of
algorithm runs performed.
}
\description{
Evaluates the candidate configurations listed on \code{configs.to.eval} on the
tuning instances listed in \code{instances.to.eval}. Notice that this routine does
not repeat evaluations - if a configuration \code{i} already has a performance
value for instance \code{j} that evaluation is skipped.
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br}),
Athila Trindade (\email{rochaathila@gmail.com})
}
