% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitLinearModel.R
\name{FitLinearModels}
\alias{FitLinearModels}
\title{Fit linear regression models}
\usage{
FitLinearModels(X, perturbed.models, model.order)
}
\arguments{
\item{X}{data frame containing the performance data}

\item{perturbed.models}{number of perturbed models to generate}

\item{model.order}{order to use for the regression model. Defaults to 2}
}
\value{
list containing information about all models fitted
}
\description{
Fits linear regression models to the performance data gathered by MetaTuner.
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br}),
Athila Trindade (\email{rochaathila@gmail.com})
}
