% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitModels.R
\name{FitModels}
\alias{FitModels}
\title{Fit regression model to performance data}
\usage{
FitModels(X, perturbed.models, model.order = 2, type = c("linear",
  "quantile"), ...)
}
\arguments{
\item{X}{data frame containing the performance data}

\item{perturbed.models}{number of perturbed models to generate}

\item{model.order}{order to use for the regression model. Defaults to 2}

\item{type}{type of model to fit. \code{linear} for linear regression using OLS,
\code{quantile} for quantile regression of the median}

\item{...}{other parameters to be passed down to specific fitting functions}
}
\value{
list containing information about all models fitted
}
\description{
Fits regression models to the performance data gathered by MetaTuner.
Returns the model fitted to the data, plus a number of models obtained by
perturbing the regression coefficients within certain error bounds.
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br}),
Athila Trindade (\email{rochaathila@gmail.com})
}
