% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitQuantileModels.R
\name{FitQuantileModels}
\alias{FitQuantileModels}
\title{Fit quantile regression models}
\usage{
FitQuantileModels(X, perturbed.models, model.order)
}
\arguments{
\item{X}{data frame containing the performance data}

\item{perturbed.models}{number of perturbed models to generate}

\item{model.order}{order to use for the regression model. Defaults to 2}
}
\value{
list containing information about all models fitted
}
\description{
Fits quantile regression models to the performance data gathered by
MetaTuner.
}
\author{
Felipe Campelo (\email{fcampelo@ufmg.br}),
Athila Trindade (\email{rochaathila@gmail.com})
}
