% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateInitialSample.R
\name{GenerateInitialSample}
\alias{GenerateInitialSample}
\title{Generates the initial sampling of the space of configurations}
\usage{
GenerateInitialSample(m0, dim, method = c("lhs", "sobol"), ndigits = 4, ...)
}
\arguments{
\item{m0}{number of points to be generated}

\item{dim}{dimension for generating the sample (i.e., number of parameters
being tuned)}

\item{method}{type of method to be used in the generation of the sample (see
section \code{Methods} for details).}

\item{ndigits}{number of decimal places to use for each parameter.}

\item{...}{further parameters to be passed down to the specific methods (see
section \code{Methods} for details).}
}
\value{
a list vector where each object is a configuration list,
containing the following fields:
- \code{config}, a named list containing parameter values
- \code{Yij}, a data frame with two columns: \code{instance.ID} and \code{y}
(the performance value of the configuration on the instance).
This data frame is generated empty.
- \code{perf}, a numeric scalar containing the summary performance value
of the configuration
}
\description{
Generates an initial sample within the space of configurations using a
space-covering design. The initial sample is scaled to the interval \code{[0,1]}
for all parameters
}
\section{Methods}{

Currently there are two methods implemented for the generation of the initial
sample:
\itemize{
\item Latin hypercube sampling ("lhs"): a space-covering design based on Latin
Hypercube Sampling. Needs no additional parameters. Needs no additional
parameters.
\item Low-discrepancy sequences of points ("sobol"): a space-covering design
using Sobol's low-discrepancy sequences of points. Based on the
implementation of package \code{SobolSequence}. Needs no additional parameters.
}
}

\author{
Felipe Campelo (\email{fcampelo@ufmg.br}),
Athila Trindade (\email{rochaathila@gmail.com})
}
